requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
psycopg[binary]
python-dotenv
pydantic
